+++ ruma-client-api/CHANGELOG.md
@@ -58,6 +58,7 @@ Breaking changes:
   the `ruma-common` crate
 * Replace `r0::room::create_room::InitialStateEvent` with `ruma_events::InitialStateEvent`
 * `error::ErrorKind` no longer implements `Copy`, `FromStr`
+* Switch from `AnyEvent` to `AnyRoomEvent` in `r0::search::search_events`
 
 Improvements: